Change BSP tree to double precision.
Single precision gave errors on saving; intersection code
is sensitive to precision.

+++ roomedit/include/bsp.h
@@ -98,18 +98,18 @@
 /* plane defined by ax + by + c = 0. (x and y are in fineness units.) */
 typedef struct
 {
-   float a, b, c;
+  double a, b, c;
 } Plane;
 
 /* box defined by its top left and bottom right coordinates (in fineness) */
 typedef struct
 {
-   float x0,y0,x1,y1;
+   double x0,y0,x1,y1;
 } Box;
 
 typedef struct
 {
-   float x,y;
+   double x,y;
 } Pnt;
 
 typedef struct WallData
@@ -133,9 +133,9 @@ typedef struct WallData
    int pos_sector;             /* Sector # on + side */
    int neg_sector;             /* Sector # on - side */
 
-   float x0, y0, x1, y1;       /* coordinates of wall start and end */
+   double x0, y0, x1, y1;       /* coordinates of wall start and end */
 
-   float length;               /* length of wall; 1 grid square = 64 */
+   double length;               /* length of wall; 1 grid square = 64 */
    int z0;                     /* height of bottom of lower wall */
    int z1;                     /* height of top of lower wall / bottom of normal wall */
    int z2;                     /* height of top of normal wall / bottom of upper wall */
